requirements.txt
requests < 3
pyyaml < 7